Music: fix the javacrash when music application monkey test
BZ: 73098

When music application do audo shuffle mode, a queue-change Intent
is sent to TraceBrowserActivity, But sometimes it has not time to handle
queue-change event when UI is tapped or clicked. Before DropListener.drop
is called, queue is checked if it's changed. If yes, manually call onReceive
to refresh NowPlayingList firstly.

Change-Id: I23745a70384d75c28c4e000fdf7d1e6ee69d3c08
Signed-off-by: Zhang Dongxing <dongxing.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/79732
Reviewed-by: Kalugin, Evgeny <evgeny.kalugin@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Wu, JunX <junx.wu@intel.com>
Reviewed-by: Zhang, Hongyu <hongyu.zhang@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>