[Google - Upstreaming]follow up needed per google's comments in upstream
BZ: 81517

The original patch was to do progressbar updates only when the
screen is ON. Used PowerManager functions in run() of the Runnable class for this.
Google proposed alternate method to do this. Handle it by setting
a flag false in onPause() and true in onResume() as whenever
the activity goes to background, onPause is invoked.

Change-Id: I544cda105cc4a7a4ed040b83d628d761dedbd81e
Signed-off-by: jpadmana <jayashree.r.padmanaban@intel.com>
Reviewed-on: http://android.intel.com:8080/125051
Reviewed-by: M, Arulselvan <arulselvan.m@intel.com>
Tested-by: M, Arulselvan <arulselvan.m@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>